refactor(tradeComplianceLibrary): 重构海关编码税则和计量单位模块
- 删除 ErpCustomsPortsController 及相关接口实现
- 海关编码税则相关路径及权限前缀改为 trade-compliance-library/hscode
- 计量单位相关路径及权限前缀改为 trade-compliance-library/units
- 统一海关编码税则、计量单位的注释描述和命名，完善提示信息
- 优化海关编码税则的查询条件构造逻辑，改进匹配方式
- 更新海关编码税则及计量单位的服务接口及实现逻辑描述
- 修改主键字段的校验信息为更友好的“ID不能为空”
- 在前端路由中新增海关编码税则和计量单位页面及对应API服务封装
- 计量单位删除标记字段类型由 Integer 修改为 Long
- 优化Excel导出字段名称以及注解描述一致性调整

+++ smart-customs-modules/business/src/main/java/io/github/yak33/business/domain/ErpUnits.java
@@ -8,10 +8,10 @@ import lombok.EqualsAndHashCode;
 import java.io.Serial;
 
 /**
- * 货物单位编码对象 erp_units
+ * 计量单位对象 erp_units
  *
  * @author ZHANGCHAO
- * @date 2025-10-18
+ * @date 2026-01-19
  */
 @Data
 @EqualsAndHashCode(callSuper = true)
@@ -38,7 +38,7 @@ public class ErpUnits extends TenantEntity {
     private String name;
 
     /**
-     * 英文代码
+     * 单位英文代码
      */
     private String itemKey;
 
@@ -51,7 +51,7 @@ public class ErpUnits extends TenantEntity {
      * 删除标记
      */
     @TableLogic
-    private Integer delFlag;
+    private Long delFlag;
 
     /**
      * 创建部门 - 数据库表中不存在此字段，排除映射
